+++ pom.xml
@@ -3,7 +3,7 @@
     <parent>
         <groupId>org.sonatype.oss</groupId>
         <artifactId>oss-parent</artifactId>
-        <version>7</version>
+        <version>9</version>
         <relativePath/>
     </parent>
 
@@ -11,7 +11,7 @@
 
     <groupId>com.github.liaochong</groupId>
     <artifactId>myexcel</artifactId>
-    <version>2.8.5</version>
+    <version>2.8.6</version>
     <packaging>jar</packaging>
 
     <name>myexcel</name>
